logica de boton delete

+++ Models/Book.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Numerics;
 using System.Threading.Tasks;
@@ -8,23 +9,19 @@ using System.Threading.Tasks;
 namespace Biblioteca_c_sharp.Models;
 
 public class Book
-{   
+{
     [Key]
-    public BigInteger Id { get; set; }
-    
+    public long Id { get; set; }
     public string Title { get; set; }
-
     public string Author { get; set; }
-
     public string Isbn { get; set; }
-
     public string Category { get; set; }
-
     [Required]
     public bool Available { get; set; }
 
-    public Book (BigInteger id, string title, string author, string isbn, string category, bool available)
-    {
+    // Constructor con parámetros públicos
+    public Book(long id ,string title, string author, string isbn, string category, bool available)
+    {   
         Id = id;
         Title = title;
         Author = author;
@@ -32,5 +29,4 @@ public class Book
         Category = category;
         Available = available;
     }
-
 }
